+++ lib/more_than_wrap.dart
@@ -1,5 +1,8 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+
+part 'implementation/render_widget.dart';
+part 'implementation/render_box.dart';
+part 'implementation/widget.dart';